% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fst2probs.R
\name{fst2probs}
\alias{fst2probs}
\title{Extract genotype probabilities from fst database}
\usage{
fst2probs(object)
}
\arguments{
\item{object}{Object of class \code{"fst_genoprob"}, linking to an fst database of genotype probabilities.}
}
\value{
An object of class \code{"calc_genoprob"} (a list of 3-dimensional arrays).
}
\description{
Extract genotype probabilities from fst database as an ordinary calc_genoprob object.
}
\details{
The genotype probabilities are extracted from the fst database. Each chromosome is extracted in turn.
}
\examples{
library(qtl2)
grav2 <- read_cross2(system.file("extdata", "grav2.zip", package="qtl2"))
map <- insert_pseudomarkers(grav2$gmap, step=1)
probs <- calc_genoprob(grav2, map, error_prob=0.002)
dir <- tempdir()
fprobs <- probs2fst(probs, "grav2", dir)
nprobs <- fst2probs(fprobs)
}
\seealso{
\code{\link[=probs2fst]{probs2fst()}}
}
\keyword{utilities}
