% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/restore_fstprobs.R
\name{fst_genoprob_restore}
\alias{fst_genoprob_restore}
\title{Restore fst_genoprob object to original dimensions.}
\usage{
fst_genoprob_restore(object)
}
\arguments{
\item{object}{Object of class \code{\link{fst_genoprob}}.}
}
\value{
Element of class \code{\link{fst_genoprob}}.
}
\description{
Any \code{\link{fst_genoprob}} object has embedded its original data and dimensions.
This resets elements \code{ind}, \code{chr} and \code{mar} to the full set.
}
\details{
Object is unclassed and elements \code{ind}, \code{chr} and \code{mar} are changed before
reseting attributes as \code{\link{fst_genoprob}} object.
See \code{\link{fst_genoprob}} return description for details.
}
\examples{
library(qtl2)
grav2 <- read_cross2(system.file("extdata", "grav2.zip", package="qtl2"))
map <- insert_pseudomarkers(grav2$gmap, step=1)
probs <- calc_genoprob(grav2, map, error_prob=0.002)
dir <- tempdir()
fprobs <- fst_genoprob(probs, "grav2", dir)
dim(fprobs)
fprobs2 <- subset(fprobs, chr=1:2)
dim(fprobs2)
fprobs5 <- fst_genoprob_restore(fprobs2)
dim(fprobs5)

}
\keyword{utilities}
