% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qtl2fst-package.R
\docType{package}
\name{qtl2fst-package}
\alias{qtl2fst}
\alias{qtl2fst-package}
\title{qtl2fst: Database Storage of Genotype Probabilities for QTL Mapping}
\description{
Uses the 'fst' package to store genotype probabilities on disk for the 'qtl2' package. These genotype probabilities are a central data object for mapping quantitative trait loci (QTL), but they can be quite large. The facilities in this package enable the genotype probabilities to be stored on disk, leading to reduced memory usage with only a modest increase in computation time.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/rqtl/qtl2fst}
}

}
\author{
\strong{Maintainer}: Karl W Broman \email{broman@wisc.edu} (\href{https://orcid.org/0000-0002-4914-6671}{ORCID})

Authors:
\itemize{
  \item Karl W Broman \email{broman@wisc.edu} (\href{https://orcid.org/0000-0002-4914-6671}{ORCID})
  \item Brian S Yandell \email{byandell@wisc.edu} (\href{https://orcid.org/0000-0002-8774-9377}{ORCID})
  \item Petr Simecek (\href{https://orcid.org/0000-0002-2922-7183}{ORCID})
}

}
\keyword{internal}
