% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rbind_fst_genoprob.R
\name{rbind.fst_genoprob}
\alias{rbind.fst_genoprob}
\title{Join genotype probabilities for different individuals}
\usage{
\method{rbind}{fst_genoprob}(..., fbase, fdir = NULL)
}
\arguments{
\item{...}{Genotype probability objects as produced by
\code{\link{fst_genoprob}}. Must have the same set of markers and
genotypes.}

\item{fbase}{Base of fileame for fst database.
Needed if objects have different fst databases.}

\item{fdir}{Directory for fst database.}
}
\value{
A single genotype probability object.
}
\description{
Join multiple genotype probability objects, as produced by
\code{\link{fst_genoprob}} for different individuals.
}
\examples{
library(qtl2)
grav2 <- read_cross2(system.file("extdata", "grav2.zip", package="qtl2"))
map <- insert_pseudomarkers(grav2$gmap, step=1)
probsA <- calc_genoprob(grav2[1:5,], map, error_prob=0.002)
probsB <- calc_genoprob(grav2[6:12,], map, error_prob=0.002)
dir <- tempdir()
fprobsA <- fst_genoprob(probsA, "exampleAr", dir)
fprobsB <- fst_genoprob(probsB, "exampleBr", dir)
fprobs <- rbind(fprobsA, fprobsB, fbase = "exampleABr")

}
