% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fst2calc_genoprob.R
\name{fst2calc_genoprob}
\alias{fst2calc_genoprob}
\title{Store genotype probabilities in fst database}
\usage{
fst2calc_genoprob(object)
}
\arguments{
\item{object}{Object of class \code{\link{fst_genoprob}}.}
}
\value{
An object of class \code{\link[qtl2]{calc_genoprob}}.
}
\description{
Uses package fst to convert R object created in R/qtl2 for fast access.
}
\details{
The genotype probabilities are extracted from 1-2 fst databases. Each chromosome is extracted in turn.
}
\examples{
library(qtl2)
grav2 <- read_cross2(system.file("extdata", "grav2.zip", package="qtl2"))
map <- insert_pseudomarkers(grav2$gmap, step=1)
probs <- calc_genoprob(grav2, map, error_prob=0.002)
dir <- tempdir()
fprobs <- fst_genoprob(probs, "grav2", dir)
nprobs <- fst2calc_genoprob(fprobs)

}
\keyword{utilities}
