% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fst_path.R
\name{replace_path}
\alias{replace_path}
\title{Replace the path used in fst_genoprob object}
\usage{
replace_path(object, path)
}
\arguments{
\item{object}{An object of class \code{"fst_genoprob"} as created by \code{\link[=probs2fst]{probs2fst()}}.}

\item{path}{New path (directory + file stem as a single character string) to be used in the object.}
}
\value{
The input \code{object} with the path replaced.
If any of the expected files don't exist with the new path, warnings are issued.
}
\description{
Replace the path used in an fst_genoprob object.
}
\seealso{
\code{\link[=fst_path]{fst_path()}}, \code{\link[=fst_files]{fst_files()}}
}
