% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.R
\name{summary.fst_genoprob}
\alias{summary.fst_genoprob}
\title{Summary of an fst_genoprob object}
\usage{
\method{summary}{fst_genoprob}(object, ...)
}
\arguments{
\item{object}{An object of class \code{"fst_genoprob"}, as output by \code{\link[=fst_genoprob]{fst_genoprob()}}.}

\item{...}{Ignored.}
}
\description{
Summarize an fst_genoprob object
}
\keyword{utilities}
